refactor: Restructure demo components and data management
- Removed legacy demo components and data files to streamline the project structure.
- Introduced new timeline data files for basic, mixed, and nested timelines, enhancing data organization.
- Updated demo components to utilize the new data structure, improving maintainability and clarity.
- Added new layout and routing components for better navigation across timeline demos.
- Adjusted tests to reflect changes in component structure and data handling.
