+++ src/demo/components/index.ts
@@ -1,0 +1,3 @@
+// Demo components exports
+export * from './vertical';
+export * from './horizontal';
