--- src/demo/components/special/index.ts
@@ -1,1 +1,0 @@
-export { DynamicLoad } from './DynamicLoad';
