--- src/demo/components/vertical/VerticalBasic.tsx
@@ -1,52 +1,0 @@
-import { TimelineItemModel } from '@models/TimelineItemModel';
-import React, { FunctionComponent } from 'react';
-import Chrono from '../../../components';
-import { ComponentContainerTree, Vertical } from '../../App.styles';
-
-export const VerticalBasic: FunctionComponent<{
-  type: string;
-  items: TimelineItemModel[];
-}> = ({ type, items }) => {
-  return (
-    <Vertical id="vertical">
-      <ComponentContainerTree type={type}>
-        <Chrono
-          items={items}
-          mode="VERTICAL"
-          slideShow
-          cardWidth={650}
-          slideItemDuration={2500}
-          scrollable={{ scrollbar: false }}
-          noUniqueId
-          uniqueId="vertical_basic_test"
-          parseDetailsAsHTML
-          highlightCardsOnHover
-          enableQuickJump={true}
-          toolbarPosition="top"
-          onItemSelected={(selected) => console.log(selected.index)}
-          fontSizes={{
-            title: '1.5rem',
-          }}
-          theme={{
-            cardDetailsColor: '#555555',
-          }}
-          focusActiveItemOnLoad
-          activeItemIndex={2}
-          cardHeight={200}
-          contentDetailsHeight={10}
-          timelinePointDimension={20}
-          classNames={{
-            cardText: 'custom-text',
-          }}
-          mediaSettings={{
-            fit: 'cover',
-          }}
-          enableDarkToggle
-          enableBreakPoint={true}
-          responsiveBreakPoint={768}
-          disableToolbar
-        />
-      </ComponentContainerTree>
-    </Vertical>
-  );
-};
