+++ src/demo/data/nested-timeline.ts
@@ -1,6 +1,6 @@
 import { TimelineItemModel } from '@models/TimelineItemModel';
 
-const items: TimelineItemModel[] = [
+const nestedTimelineData: TimelineItemModel[] = [
   {
     title: 'May 1940',
     cardTitle: 'Dunkirk',
@@ -9,7 +9,6 @@ const items: TimelineItemModel[] = [
       name: 'dunkirk beach',
       source: {
         url: 'https://i2-prod.mirror.co.uk/incoming/article10847802.ece/ALTERNATES/s810/PAY-Dunkirk-in-colour.jpg',
-        // url: "/dunkirk.mp4"
       },
       type: 'IMAGE',
     },
@@ -21,20 +20,6 @@ const items: TimelineItemModel[] = [
     title: '25 July 1940',
     cardTitle: 'The Battle of Britain',
     items: [
-      // {
-      //   cardTitle: 'May 10, 1940',
-      //   cardDetailedText:
-      //     'Germany invades France, Belgium, and the Netherlands',
-      //   url: 'http://www.google.com',
-      //   // media: {
-      //   //   name: 'dunkirk beach',
-      //   //   source: {
-      //   //     url: 'https://i2-prod.mirror.co.uk/incoming/article10847802.ece/ALTERNATES/s810/PAY-Dunkirk-in-colour.jpg',
-      //   //     // url: "/dunkirk.mp4"
-      //   //   },
-      //   //   type: 'IMAGE',
-      //   // },
-      // },
       {
         cardTitle: 'June 18, 1940',
         url: 'http://www.google.com',
@@ -46,19 +31,9 @@ const items: TimelineItemModel[] = [
           'The Germans launch "Adler Tag" (Eagle Day), a massive assault on British airfields and radar stations. The Luftwaffe hopes to achieve a decisive victory and gain control of the skies over Britain, but the RAF is able to mount a strong defense and inflict heavy losses on the Germans',
         url: 'http://www.google.com',
       },
-      // {
-      //   cardTitle: 'August 15, 1940',
-      //   cardDetailedText: `The Luftwaffe mistakenly bombs London, prompting Churchill to order retaliatory attacks on Berlin. The bombing of London marks a significant shift in the German strategy, as the focus shifts from attacking airfields to targeting cities and civilian populations`,
-      //   url: 'http://www.google.com',
-      // },
-      // {
-      //   cardTitle: 'August 24, 1940',
-      //   cardDetailedText: ` The Germans launch a second wave of attacks on British airfields, known as "The Hardest Day." The attacks result in the heaviest losses of the battle for the RAF, but the Germans also suffer significant losses.`,
-      //   url: 'http://www.google.com',
-      // },
     ],
     cardSubtitle: `RAF Spitfire pilots scramble for their planes`,
-    cardDetailedText: `After France’s surrender in June 1940, Churchill told the British people, “Hitler knows that he will have to break us in this island or lose the war”. To mount a successful invasion, the Germans had to gain air superiority. The first phase of the battle began on 10 July with Luftwaffe attacks on shipping in the Channel.
+    cardDetailedText: `After France's surrender in June 1940, Churchill told the British people, "Hitler knows that he will have to break us in this island or lose the war". To mount a successful invasion, the Germans had to gain air superiority. The first phase of the battle began on 10 July with Luftwaffe attacks on shipping in the Channel.
       The following month, RAF Fighter Command airfields and aircraft factories came under attack. Under the dynamic direction of Lord Beaverbrook, production of Spitfire and Hurricane fighters increased, and despite its losses in pilots and planes, the RAF was never as seriously weakened as the Germans supposed.`,
   },
   {
@@ -71,9 +46,9 @@ const items: TimelineItemModel[] = [
       type: 'VIDEO',
     },
     cardSubtitle: `A column of Red Army prisoners taken during the first days of the German invasion`,
-    cardDetailedText: `Since the 1920s, Hitler had seen Russia, with its immense natural resources, as the principal target for conquest and expansion. It would provide, he believed, the necessary ‘Lebensraum’, or living space, for the German people. And by conquering Russia, Hitler would also destroy the “Jewish pestilential creed of Bolshevism”. His non-aggression pact with Stalin in August 1939 he regarded as a mere temporary expedient.
+    cardDetailedText: `Since the 1920s, Hitler had seen Russia, with its immense natural resources, as the principal target for conquest and expansion. It would provide, he believed, the necessary 'Lebensraum', or living space, for the German people. And by conquering Russia, Hitler would also destroy the "Jewish pestilential creed of Bolshevism". His non-aggression pact with Stalin in August 1939 he regarded as a mere temporary expedient.
       Barely a month after the fall of France, and while the Battle of Britain was being fought, Hitler started planning for the Blitzkrieg campaign against Russia, which began on 22 June 1941. Despite repeated warnings, Stalin was taken by surprise, and for the first few months the Germans achieved spectacular victories, capturing huge swathes of land and hundreds of thousands of prisoners. But they failed to take Moscow or Leningrad before winter set in.
-      On 5/6 December, the Red Army launched a counter-offensive which removed the immediate threat to the Soviet capital. It also brought the German high command to the brink of a catastrophic military crisis. Hitler stepped in and took personal command. His intervention was decisive and he later boasted, “That we overcame this winter and are today in a position again to proceed victoriously… is solely attributable to the bravery of the soldiers at the front and my firm will to hold out…”`,
+      On 5/6 December, the Red Army launched a counter-offensive which removed the immediate threat to the Soviet capital. It also brought the German high command to the brink of a catastrophic military crisis. Hitler stepped in and took personal command. His intervention was decisive and he later boasted, "That we overcame this winter and are today in a position again to proceed victoriously… is solely attributable to the bravery of the soldiers at the front and my firm will to hold out…"`,
   },
   {
     title: '7 December 1941',
@@ -81,13 +56,12 @@ const items: TimelineItemModel[] = [
     cardSubtitle: `The destroyer USS Shaw explodes in dry dock after being hit by Japanese aircraft`,
     media: {
       source: {
-        // url: "https://pearlharborwarbirds.com/wp-content/uploads/2016/09/Awesome-Color-Photos-of-the-Attack-on-Pearl-Harbor-1.jpg",
         url: '/pearl-harbor.mp4',
       },
       type: 'VIDEO',
       name: 'Pearl Harbor',
     },
-    cardDetailedText: `After Japan’s occupation of French Indo-China in July 1941, US President Franklin D Roosevelt, followed by Britain and the Netherlands, ordered the freezing of Japanese assets.
+    cardDetailedText: `After Japan's occupation of French Indo-China in July 1941, US President Franklin D Roosevelt, followed by Britain and the Netherlands, ordered the freezing of Japanese assets.
       Many Japanese now believed that there was no alternative between economic ruin and going to war with the United States and the European colonial powers. In October 1941, a hardline government under General Hideki Tojo came to power, and preparations were made to deliver a devastating blow against the Americans.`,
   },
   {
@@ -100,11 +74,11 @@ const items: TimelineItemModel[] = [
       },
       name: 'Fall of singapore',
     },
-    cardSubtitle: `Lieutenant General Arthur Percival and staff on their way to the Singapore Ford factory to negotiate the island’s surrender with General Yamashita`,
+    cardSubtitle: `Lieutenant General Arthur Percival and staff on their way to the Singapore Ford factory to negotiate the island's surrender with General Yamashita`,
     cardDetailedText: `The Japanese began their invasion of Malaya on 8 December 1941, and very soon the British and empire defenders were in full retreat.
-      Told previously that the Japanese were no match for European troops, morale among the defending forces slumped as General Tomoyuki Yamashita’s forces moved swiftly southwards towards Singapore.
+      Told previously that the Japanese were no match for European troops, morale among the defending forces slumped as General Tomoyuki Yamashita's forces moved swiftly southwards towards Singapore.
       The sinking of the British capital ships HMS Prince of Wales and Repulse by Japanese aircraft also contributed to the decline in morale, and panic began to set in among the civil population and the fighting troops. British commander Lieutenant General Arthur Percival had hoped to make a stand at Johore, but was forced to withdraw to Singapore Island. The Japanese landed there on 8/9 February, and before long the defence collapsed. To avoid further bloodshed, and with his water supply gone, Percival surrendered on 15 February.
-      Churchill described the surrender as, “the worst disaster… in British military history”. Over 130,000 British and empire troops surrendered to a much smaller Japanese force, which only suffered 9,824 battle casualties during the 70-day campaign. Singapore was not only a humiliating military defeat, but also a tremendous blow to the prestige of the ‘white man’ throughout Asia.`,
+      Churchill described the surrender as, "the worst disaster… in British military history". Over 130,000 British and empire troops surrendered to a much smaller Japanese force, which only suffered 9,824 battle casualties during the 70-day campaign. Singapore was not only a humiliating military defeat, but also a tremendous blow to the prestige of the 'white man' throughout Asia.`,
   },
   {
     title: '4 June 1942',
@@ -112,7 +86,7 @@ const items: TimelineItemModel[] = [
     cardSubtitle: `The American aircraft carrier USS Yorktown under Japanese attack during the battle of Midway`,
     cardDetailedText: `For six months after Pearl Harbor, just as Admiral Yamamoto predicted, Japanese forces carried all before them, capturing Hong Kong, Malaya, the Philippines and the Dutch East Indies. In May 1942, in an attempt to consolidate their grip on their new conquests, the Japanese sought to eliminate the United States as a strategic Pacific power.
       This would be done by luring into a trap the US navy carriers that had escaped Pearl Harbor, while at the same time the Japanese would occupy the Midway atoll in preparation for further attacks. The loss of the carriers would, the Japanese hoped, force the Americans to the negotiating table. In the event, it was the Americans who inflicted a crushing defeat on the Japanese. Their codebreakers were able to determine the location and date of the Japanese attack. This enabled US admiral Chester Nimitz to organise a trap of his own.
-      During the ensuing battle the Japanese suffered the loss of four carriers, one heavy cruiser and 248 aircraft, while American losses totalled one carrier, one destroyer and 98 planes. By their victory at Midway, the turning point of the Pacific war, the Americans were able to seize the strategic initiative from the Japanese, who had suffered irreplaceable losses. Admiral Nimitz described the battle’s success as “Essentially a victory of intelligence”, while President Roosevelt called it “Our most important victory in 1942… there we stopped the Japanese offensive.”`,
+      During the ensuing battle the Japanese suffered the loss of four carriers, one heavy cruiser and 248 aircraft, while American losses totalled one carrier, one destroyer and 98 planes. By their victory at Midway, the turning point of the Pacific war, the Americans were able to seize the strategic initiative from the Japanese, who had suffered irreplaceable losses. Admiral Nimitz described the battle's success as "Essentially a victory of intelligence", while President Roosevelt called it "Our most important victory in 1942… there we stopped the Japanese offensive."`,
   },
   {
     title: '25 October 1942',
@@ -124,18 +98,18 @@ const items: TimelineItemModel[] = [
         url: 'https://i.dailymail.co.uk/i/newpix/2018/03/08/11/49FEAE2A00000578-5477117-image-a-79_1520509429888.jpg',
       },
     },
-    cardDetailedText: `The North African campaign began in September1940, and for the next two years the fighting was marked by a succession of Allied and Axis advances and retreats. In the summer of 1942, the Axis forces under ‘Desert Fox’ field marshal, Erwin Rommel, looked poised to take Cairo and advance on the Suez Canal.
+    cardDetailedText: `The North African campaign began in September1940, and for the next two years the fighting was marked by a succession of Allied and Axis advances and retreats. In the summer of 1942, the Axis forces under 'Desert Fox' field marshal, Erwin Rommel, looked poised to take Cairo and advance on the Suez Canal.
       The British Middle East commander General Claude Auchinleck took personal command of the defending Eighth Army and halted the retreat at the strong defensive line at El Alamein. But Churchill, dissatisfied with Auchinleck, replaced him in August with General Harold Alexander, while Lieutenant -General Bernard Montgomery took over command of the Eighth Army.
       Montgomery immediately began to build up an enormous superiority in men and equipment, finally launching his offensive at Alamein on 23 October 1942. By the beginning of November, the Axis forces were in full retreat, although final victory in North Africa was not achieved until May 1943.
-      Although Montgomery has been criticised for being too cautious in exploiting his success at Alamein, it made him a household name and he became Britain’s most popular general of the war. Churchill hailed Alamein as a “Glorious and decisive victory… the bright gleam has caught the helmets of our soldiers, and warmed and cheered all our hearts”.`,
+      Although Montgomery has been criticised for being too cautious in exploiting his success at Alamein, it made him a household name and he became Britain's most popular general of the war. Churchill hailed Alamein as a "Glorious and decisive victory… the bright gleam has caught the helmets of our soldiers, and warmed and cheered all our hearts".`,
   },
   {
     title: 'February 1943',
     cardTitle: 'Stalingrad',
     cardSubtitle: `Red Army soldiers hoist the Soviet flag over a recaptured Stalingrad factory following the German surrender`,
-    cardDetailedText: `Throughout September and October, under General Vassili Chuikov, the city’s defenders contested every yard of ground of the devastated city.
-      The Red Army’s stubborn defence allowed General Georgi Zhukov time to prepare a counterattack that was launched on 19 November 1942, and which soon trapped the Sixth Army commanded by General Friederich Paulus.
-      Hitler, wrongly assured by Göring that the Luftwaffe could supply Stalingrad by air, ordered Paulus to hold out. He also ordered Field Marshal Erich Manstein to break through and relieve the beleaguered Sixth Army. Manstein was unsuccessful, and on 31 January 1943 Paulus capitulated. Of the 91,000 German troops who went into captivity, less than 6,000 returned home after the war. Stalingrad was one of Germany’s greatest defeats, and it effectively marked the end of Hitler’s dreams of an empire in the east.
+    cardDetailedText: `Throughout September and October, under General Vassili Chuikov, the city's defenders contested every yard of ground of the devastated city.
+      The Red Army's stubborn defence allowed General Georgi Zhukov time to prepare a counterattack that was launched on 19 November 1942, and which soon trapped the Sixth Army commanded by General Friederich Paulus.
+      Hitler, wrongly assured by Göring that the Luftwaffe could supply Stalingrad by air, ordered Paulus to hold out. He also ordered Field Marshal Erich Manstein to break through and relieve the beleaguered Sixth Army. Manstein was unsuccessful, and on 31 January 1943 Paulus capitulated. Of the 91,000 German troops who went into captivity, less than 6,000 returned home after the war. Stalingrad was one of Germany's greatest defeats, and it effectively marked the end of Hitler's dreams of an empire in the east.
       `,
   },
   {
@@ -145,7 +119,6 @@ const items: TimelineItemModel[] = [
     media: {
       type: 'VIDEO',
       source: {
-        // url: "https://www.britishpoles.uk/wp-content/uploads/2019/06/D-Day-landings-1.jpg"
         url: '/d-day.mp4',
         type: 'mp4',
       },
@@ -160,7 +133,7 @@ const items: TimelineItemModel[] = [
     cardTitle: 'The Big Three',
     cardSubtitle: `Churchill, Roosevelt and Stalin sit for a group photograph during the Yalta conference`,
     cardDetailedText: `Between June 1940 and June 1941, Britain stood alone against Hitler. But then, after the German invasion of Russia and the Japanese attack on Pearl Harbor, she gained two powerful allies.
-      For the next four years Churchill did his utmost to foster ‘The Grand Alliance’ against the Nazis. He even earned the grudging admiration of Nazi propaganda chief Dr Goebbels who said, “…I can feel only respect for this man, for whom no humiliation is too base and no trouble too great when the victory of the Allies is at stake”.
+      For the next four years Churchill did his utmost to foster 'The Grand Alliance' against the Nazis. He even earned the grudging admiration of Nazi propaganda chief Dr Goebbels who said, "…I can feel only respect for this man, for whom no humiliation is too base and no trouble too great when the victory of the Allies is at stake".
       Churchill conferred with both Roosevelt and Stalin to hammer out strategy and to discuss postwar arrangements. The three men congregated for the first time at Tehran in November 1943. There, and again at their last meeting at Yalta, Churchill was conscious of the fact that Britain, exhausted by her war effort, was now very much the junior partner of the two emerging superpowers.`,
   },
   {
@@ -168,7 +141,7 @@ const items: TimelineItemModel[] = [
     cardTitle: 'Dresden',
     cardSubtitle: `Dresden under incendiary bomb attack`,
     cardDetailedText: `At Yalta, an Allied plan to bomb the hitherto untouched city of Dresden was discussed. The reason for attacking the city was due principally to its strategic importance as a communications centre in the rear of the German retreat that followed the Soviet winter offensive of January 1945. It was also believed that Dresden might be used as an alternative to Berlin as the Reich capital.
-      The attack was part of a plan codenamed ‘Thunderclap’, designed to convince the Germans that the war was lost. It was drawn up in January 1945, when Hitler’s Ardennes offensive, V2 rocket attacks on Britain and the deployment of snorkel-equipped U-boats clearly demonstrated that Germany was still capable of offering stubborn resistance. Strategic bombing attacks had previously failed to break Germany, although they had proved valuable in reducing its capacity to wage war.
+      The attack was part of a plan codenamed 'Thunderclap', designed to convince the Germans that the war was lost. It was drawn up in January 1945, when Hitler's Ardennes offensive, V2 rocket attacks on Britain and the deployment of snorkel-equipped U-boats clearly demonstrated that Germany was still capable of offering stubborn resistance. Strategic bombing attacks had previously failed to break Germany, although they had proved valuable in reducing its capacity to wage war.
       Now, on the night of 13/14 February 1945, Dresden was attacked by 800 RAF bombers, followed by 400 bombers of the United States Army Air Force. The bombing created a firestorm that destroyed 1,600 acres of Dresden. Even today it is still uncertain as to how many died and estimates have ranged from 25,000 to 135,000. Most authorities now put the death toll at around 35,000. The scale of destruction, the enormous death toll, and its timing at such a late stage in the war, have all ensured that the bombing of Dresden still remains highly controversial.`,
   },
   {
@@ -182,7 +155,7 @@ const items: TimelineItemModel[] = [
     },
     cardSubtitle: `millions of people rejoice in the news that Germany has surrendered – the war in Europe was finally over`,
     cardDetailedText: `On the afternoon of 8 May 1945, the British prime minister Winston Churchill made the radio announcement that the world had long been waiting for.
-      “Yesterday morning,” he declared, “at 2.41 a.m., at General Eisenhower’s headquarters, General Jodl, the representative of the German High Command, and Grand Admiral Dönitz, the designated head of the German State, signed the act of unconditional surrender of all German land, sea and air forces in Europe.”
+      "Yesterday morning," he declared, "at 2.41 a.m., at General Eisenhower's headquarters, General Jodl, the representative of the German High Command, and Grand Admiral Dönitz, the designated head of the German State, signed the act of unconditional surrender of all German land, sea and air forces in Europe."
       After nearly six years, the war in Europe was finally over.`,
   },
   {
@@ -190,11 +163,11 @@ const items: TimelineItemModel[] = [
     cardTitle: 'Nagasaki',
     cardSubtitle: `Atomic bomb mushroom cloud over the Japanese city of Nagasaki`,
     cardDetailedText: `The Second World War began at dawn on Friday 1 September 1939, when Adolf Hitler launched his invasion of Poland.
-      The Poles fought bravely, but they were heavily outnumbered in both men and machines, and especially in the air. Britain and France declared war on Germany on 3 September 1939, but gave no real assistance to Poland. Two weeks later, Stalin invaded eastern Poland, and on 27 September Warsaw surrendered. Organised Polish resistance ceased after another week’s fighting. Poland was divided up between Hitler and Stalin.
+      The Poles fought bravely, but they were heavily outnumbered in both men and machines, and especially in the air. Britain and France declared war on Germany on 3 September 1939, but gave no real assistance to Poland. Two weeks later, Stalin invaded eastern Poland, and on 27 September Warsaw surrendered. Organised Polish resistance ceased after another week's fighting. Poland was divided up between Hitler and Stalin.
       In Poland the Nazis unleashed a reign of terror that was eventually to claim six million victims, half of whom were Polish Jews murdered in extermination camps. The Soviet regime was no less harsh. In March and April 1940, Stalin ordered the murder of over 20,000 Polish officers and others who had been captured in September 1939. Tens of thousands of Poles were also forcibly deported to Siberia.
       By May 1945, and despite his promises to Churchill and Roosevelt, Stalin had installed a subservient communist regime in Poland.
-      Back in 1939, Poland’s then-leader Marshal Eduard Smigly-Rydz had warned, “With the Germans we risk losing our liberty, but with the Russians we lose our soul.”`,
+      Back in 1939, Poland's then-leader Marshal Eduard Smigly-Rydz had warned, "With the Germans we risk losing our liberty, but with the Russians we lose our soul."`,
   },
 ];
 
-export default items;
+export const nestedTimeline = nestedTimelineData;
