--- src/demo/horizontal-samples.tsx
@@ -1,171 +1,0 @@
-import { TimelineItemModel } from '@models/TimelineItemModel';
-import React, { FunctionComponent, useState } from 'react';
-import Chrono from '../components';
-import {
-  ComponentContainer,
-  ComponentContainerTree,
-  Horizontal,
-  Vertical,
-} from './App.styles';
-import data from './data';
-
-export const HorizontalBasic: React.FunctionComponent<{
-  type: string;
-  items: TimelineItemModel[];
-}> = ({ items }) => {
-  const [itemSelected, setItemSelected] = useState(0);
-
-  return (
-    <Horizontal id="horizontal">
-      <span>{itemSelected}</span>
-      <ComponentContainer type={'big-screen'}>
-        <Chrono
-          items={data}
-          mode="HORIZONTAL"
-          cardHeight={450}
-          cardWidth={550}
-          mediaHeight={300}
-          slideShow
-          slideItemDuration={2550}
-          itemWidth={300}
-          onItemSelected={(selected) => setItemSelected(selected.index)}
-          timelinePointDimension={20}
-          timelinePointShape="square"
-          // disableInteraction
-          // cardPositionHorizontal="TOP"
-          parseDetailsAsHTML
-          buttonTexts={{
-            first: 'Jump to First',
-            last: 'Jump to Last',
-            next: 'Next',
-            previous: 'Previous',
-          }}
-          enableDarkToggle
-          mediaSettings={{
-            align: 'center',
-            fit: 'cover',
-          }}
-        >
-          <div className="chrono-icons">
-            <img src="color-circle.svg" alt="github" />
-            <img src="color-circle.svg" alt="github" />
-            <img src="color-circle.svg" alt="github" />
-            <img src="color-circle.svg" alt="github" />
-            <img src="color-circle.svg" alt="github" />
-          </div>
-        </Chrono>
-      </ComponentContainer>
-    </Horizontal>
-  );
-};
-
-export const HorizontalAll: React.FunctionComponent<{
-  type: string;
-  items: TimelineItemModel[];
-}> = ({ items }) => {
-  const [index, setIndex] = React.useState(-1);
-  return (
-    <Horizontal id="horizontal">
-      <ComponentContainer type={'big-screen'}>
-        <span>{index}</span>
-        <Chrono
-          items={items}
-          mode="HORIZONTAL"
-          cardHeight={350}
-          cardWidth={500}
-          enableDarkToggle
-          slideShow
-          textOverlay
-          mediaHeight={250}
-          slideItemDuration={2550}
-          itemWidth={400}
-          focusActiveItemOnLoad
-          onItemSelected={(selected) => {
-            setIndex(selected.index);
-          }}
-          timelinePointDimension={20}
-          showAllCardsHorizontal
-          activeItemIndex={8}
-        >
-          <div className="chrono-icons">
-            <img src="color-circle.svg" alt="github" />
-            <img src="color-circle.svg" alt="github" />
-            <img src="color-circle.svg" alt="github" />
-            <img src="color-circle.svg" alt="github" />
-            <img src="color-circle.svg" alt="github" />
-          </div>
-        </Chrono>
-      </ComponentContainer>
-    </Horizontal>
-  );
-};
-
-export const HorizontalInitialSelectedItem: React.FunctionComponent<{
-  type: string;
-  items: TimelineItemModel[];
-}> = ({ items }) => {
-  return (
-    <Horizontal id="horizontal">
-      <ComponentContainer type={'big-screen'}>
-        <Chrono
-          items={data}
-          activeItemIndex={2}
-          mode="HORIZONTAL"
-          cardHeight={150}
-          slideShow
-          slideItemDuration={2550}
-          itemWidth={200}
-          cardLess
-          onItemSelected={(selected) => console.log(selected)}
-        >
-          <div className="chrono-icons">
-            <img src="color-circle.svg" alt="github" />
-            <img src="color-circle.svg" alt="github" />
-            <img src="color-circle.svg" alt="github" />
-            <img src="color-circle.svg" alt="github" />
-            <img src="color-circle.svg" alt="github" />
-          </div>
-        </Chrono>
-      </ComponentContainer>
-    </Horizontal>
-  );
-};
-
-export const HorizontalBasicCardLess: FunctionComponent<{
-  type: string;
-  items: TimelineItemModel[];
-}> = ({ type, items }) => (
-  <Vertical id="vertical">
-    <ComponentContainerTree type={type}>
-      <Chrono
-        items={items}
-        mode="HORIZONTAL"
-        cardLess
-        theme={{
-          cardBgColor: '#fff',
-          titleColorActive: 'red',
-        }}
-        onItemSelected={(selected) => console.log(selected.cardTitle)}
-      />
-    </ComponentContainerTree>
-  </Vertical>
-);
-
-export const HorizontalSlideshow: FunctionComponent<{
-  type: string;
-  cardHeight?: number;
-  items: TimelineItemModel[];
-}> = ({ type, cardHeight }) => (
-  <Horizontal id="slideshow">
-    <ComponentContainer type={type}>
-      <Chrono
-        items={data}
-        mode="HORIZONTAL"
-        slideShow
-        slideItemDuration={1500}
-        cardHeight={450}
-        scrollable
-      />
-    </ComponentContainer>
-  </Horizontal>
-);
