--- src/demo/layout/Layout.module.scss
@@ -1,155 +1,0 @@
-.wrapper {
-  display: flex;
-  flex-direction: column;
-  min-height: 100vh;
-  background: #f8fafc;
-  width: 100%;
-}
-
-.header {
-  flex-shrink: 0;
-  border-radius: 0;
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-}
-
-.container {
-  min-height: 100vh;
-  display: flex;
-  background: linear-gradient(135deg, #f0f4ff 0%, #e6f0ff 100%);
-}
-
-.navigation {
-  width: 280px;
-  background: rgba(255, 255, 255, 0.1);
-  backdrop-filter: blur(10px);
-  border-right: 1px solid rgba(255, 255, 255, 0.2);
-  padding: 2rem 1.5rem;
-  box-shadow: 0 8px 32px rgba(31, 38, 135, 0.37);
-}
-
-.logo {
-  font-size: 1.75rem;
-  font-weight: 700;
-  margin-bottom: 2rem;
-  color: white;
-  text-shadow: 0 2px 4px rgba(0, 0, 0, 0.3);
-}
-
-.navSection {
-  margin-bottom: 2rem;
-}
-
-.navSectionTitle {
-  font-size: 0.875rem;
-  font-weight: 600;
-  color: rgba(255, 255, 255, 0.8);
-  text-transform: uppercase;
-  letter-spacing: 0.5px;
-  margin-bottom: 0.75rem;
-}
-
-.navList {
-  list-style: none;
-  padding: 0;
-  margin: 0;
-}
-
-.navItem {
-  margin-bottom: 0.25rem;
-}
-
-.navLink {
-  display: block;
-  padding: 0.75rem 1rem;
-  color: rgba(255, 255, 255, 0.9);
-  text-decoration: none;
-  border-radius: 8px;
-  transition: all 0.2s ease;
-  font-weight: 500;
-  
-  &:hover {
-    background: rgba(255, 255, 255, 0.1);
-    color: white;
-    transform: translateX(4px);
-  }
-  
-  &.active {
-    background: rgba(255, 255, 255, 0.2);
-    color: white;
-    box-shadow: 0 4px 12px rgba(0, 0, 0, 0.15);
-  }
-}
-
-.content {
-  flex: 1;
-  padding: 2rem;
-  overflow-y: auto;
-  background: white;
-  margin: 1rem;
-  border-radius: 16px;
-  box-shadow: 0 10px 40px rgba(0, 0, 0, 0.1);
-}
-
-.mobileNav {
-  display: none;
-}
-
-@media (max-width: 768px) {
-  .container {
-    flex-direction: column;
-  }
-  
-  .navigation {
-    display: none;
-    width: 100%;
-    position: fixed;
-    top: 0;
-    left: 0;
-    z-index: 1000;
-    height: 100vh;
-    
-    &.mobileOpen {
-      display: block;
-    }
-  }
-  
-  .mobileNav {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    padding: 1rem;
-    background: rgba(255, 255, 255, 0.95);
-    backdrop-filter: blur(10px);
-    border-bottom: 1px solid rgba(0, 0, 0, 0.1);
-    position: sticky;
-    top: 0;
-    z-index: 100;
-  }
-  
-  .mobileTitle {
-    font-size: 1.25rem;
-    font-weight: 600;
-    color: #667eea;
-  }
-  
-  .mobileMenuButton {
-    background: none;
-    border: none;
-    font-size: 1.5rem;
-    color: #667eea;
-    cursor: pointer;
-    padding: 0.5rem;
-    border-radius: 8px;
-    transition: background 0.2s ease;
-    
-    &:hover {
-      background: rgba(102, 126, 234, 0.1);
-    }
-  }
-  
-  .content {
-    margin: 0;
-    border-radius: 0;
-    padding: 1rem;
-  }
-}
