+++ types/static.d.ts
@@ -2,6 +2,10 @@
 /* Use this folder to also add/extend a package d.ts file, if needed. */
 
 declare module '*.css';
+declare module '*.scss' {
+  const classes: { [key: string]: string };
+  export default classes;
+}
 declare module '*.svg' {
   const ref: string;
   export default ref;
